Added the Choice model
The choice model has a title and belongs to both a User and a
Decision. Create Choice objects through the Decision object on
a User. The creator property is the association with the User
who created the Choice. If no creator is set at save, the creator
defaults to the creator of the Decision.