Lots of improvements.
The app now downloads all the user's decisions. Does not download the decisions the user is participating in though. The web app has a small bug with generating the JSON for a decision. The my_votes attribute always returns false for the iOS app. The user can also vote on decisions.