Handle timezones in the user's timezone.
Previously, we were just using utc for everything. This meant the
user had to convert his/her deadline to utc when entereing it. This
is unacceptable. Instead, make it so the user can enter in deadlines
in his/her timezone, and then convert it to utc on the server. Then
display it to users viewing the decision in their local timezone.